requirement.txt
pytesseract
opencv-python
pillow
PyQt5